% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_expanded_array.R
\name{bind_expanded_array}
\alias{bind_expanded_array}
\title{Bind two expanded arrays}
\usage{
bind_expanded_array(exparr1, exparr2, axis)
}
\arguments{
\item{exparr1}{An expanded array}

\item{exparr2}{An expanded array}

\item{axis}{An axis to bind}
}
\description{
Bind two expanded arrays
}
