% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplot_PCA.R
\name{biplot_PCA}
\alias{biplot_PCA}
\title{Plot results of prcomp() and PPCA()}
\usage{
biplot_PCA(
  pca_res,
  x_pc = 1,
  y_pc = 2,
  labels = NULL,
  colors = NULL,
  filename = NULL
)
}
\arguments{
\item{pca_res}{Results of PCA (`prcomp()` or `PPCA()`).}

\item{x_pc}{A numeric vector of PCs to be used as x-axis.}

\item{y_pc}{A numeric vector of PCs to be used as y-axis. Its length should be the same as `x_pc`.}

\item{labels}{A character vector to be shown as hover text in plotly.}

\item{colors}{A character vector of groups which is reflected to point colors.}

\item{filename}{A character to be used in filename.}
}
\description{
Plot results of prcomp() and PPCA()
}
