% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpandedArray.R
\name{call_expanded_array}
\alias{call_expanded_array}
\title{An expanded class of array}
\usage{
call_expanded_array()
}
\arguments{
\item{data}{A vector, a matrix or an array.}

\item{dim}{Dimension of data.}

\item{diminfo}{A list of information of the array.}
}
\description{
An expanded class of array
}
