% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind_expanded_array.R
\name{cbind_expanded_array}
\alias{cbind_expanded_array}
\title{Bind two expanded arrays in column direction}
\usage{
cbind_expanded_array(exparray1, exparray2)
}
\arguments{
\item{exparray1}{An expanded array.}

\item{exparray2}{An expanded array.}
}
\description{
Bind two expanded arrays in column direction
}
