% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbind_exparray.R
\name{cbind_exparray}
\alias{cbind_exparray}
\title{Bind two expanded arrays in a column direction}
\usage{
cbind_exparray(exparr1, exparr2)
}
\arguments{
\item{exparr1}{An expanded array}

\item{exparr2}{An expanded array}
}
\description{
Bind two expanded arrays in a column direction
}
