% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir_create_with_file_path.R
\name{dir_create_with_file_path}
\alias{dir_create_with_file_path}
\title{Create directory and output its path}
\usage{
dir_create_with_file_path(...)
}
\arguments{
\item{...}{Arguments of `file.path()`.}
}
\description{
Create directory and output its path
}
