% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_array.R
\name{empty_array}
\alias{empty_array}
\title{Create empty matrix with dimnames}
\usage{
empty_array(..., data = NA)
}
\arguments{
\item{...}{dimnames.}

\item{data}{Data for array.}
}
\description{
Create empty matrix with dimnames
}
