% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_expanded_array.R
\name{empty_expanded_array}
\alias{empty_expanded_array}
\title{Create an empty expanded array with diminfo only}
\usage{
empty_expanded_array(diminfo)
}
\arguments{
\item{diminfo}{A list of information of the array.}
}
\description{
Create an empty expanded array with diminfo only
}
