% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_exparray.R
\name{empty_exparray}
\alias{empty_exparray}
\title{Create an empty expanded array with diminfo only}
\usage{
empty_exparray(diminfo)
}
\arguments{
\item{diminfo}{A list of information of the array.}
}
\description{
Create an empty expanded array with diminfo only
}
