% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpandedArray.R
\name{expanded_array}
\alias{expanded_array}
\title{An expanded class of array}
\usage{
expanded_array(data = NA, dim = length(data), diminfo)
}
\arguments{
\item{diminfo}{A list of information of the array.}
}
\description{
An expanded class of array
}
