% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exparray.R
\name{exparray}
\alias{exparray}
\title{Create an expanded array with data and diminfo}
\usage{
exparray(data, diminfo)
}
\arguments{
\item{data}{An array.}

\item{diminfo}{A list of information of the array.}
}
\description{
Create an expanded array with data and diminfo
}
