% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_rest_time.R
\name{print_rest_time}
\alias{print_rest_time}
\title{Print rest processing time in for loop}
\usage{
print_rest_time(iter, n_iter, start_time, n_print)
}
\arguments{
\item{iter}{Number of processed iterations.}

\item{n_iter}{Number of all iterations.}

\item{start_time}{Starting time (`proc.time()`)}

\item{n_print}{An integer. How frequently should the message be shown?}
}
\description{
Print rest processing time in for loop
}
