% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_expanded_array.R
\name{rbind_expanded_array}
\alias{rbind_expanded_array}
\title{Bind two expanded arrays in row direction}
\usage{
rbind_expanded_array(exparray1, exparray2)
}
\arguments{
\item{exparray1}{An expanded array.}

\item{exparray2}{An expanded array.}
}
\description{
Bind two expanded arrays in row direction
}
