% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_exparray.R
\name{rbind_exparray}
\alias{rbind_exparray}
\title{Bind two expanded arrays in a row direction}
\usage{
rbind_exparray(exparr1, exparr2)
}
\arguments{
\item{exparr1}{An expanded array}

\item{exparr2}{An expanded array}
}
\description{
Bind two expanded arrays in a row direction
}
