% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_data_frame.R
\name{rep_data_frame}
\alias{rep_data_frame}
\title{Replicate a data frame to row direction}
\usage{
rep_data_frame(df, n)
}
\arguments{
\item{df}{Data frame.}

\item{n}{Number of replications.}
}
\description{
Replicate a data frame to row direction
}
