% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit.select.R
\name{strsplit.select}
\alias{strsplit.select}
\title{strsplit() and select splitted elements with index}
\usage{
strsplit.select(x, split, index, from.tail = FALSE)
}
\arguments{
\item{x}{The same as strsplit.}

\item{split}{The same as strsplit.}

\item{index}{An index to be selected.}

\item{from.tail}{Weather index should be counted from tail or not.}
}
\description{
strsplit() and select splitted elements with index
}
