% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xaxis_with_vert_label.R
\name{xaxis_with_vert_label}
\alias{xaxis_with_vert_label}
\title{Add x-axis with vertical labels}
\usage{
xaxis_with_vert_label(at, labels, srt = 90, pos = 0.05)
}
\arguments{
\item{at}{Equivalent to `at` argument in `axis()` function.}

\item{labels}{Equivalent to `labels` argument in `axis()` function.}

\item{srt}{An angle for label rotation.}

\item{pos}{A parameter to adjust vertical positions of labels.}
}
\description{
Add x-axis with vertical labels
}
