% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expanded_array.R
\name{expanded_array}
\alias{expanded_array}
\title{Create an expanded array with data and diminfo}
\usage{
expanded_array(data, diminfo)
}
\arguments{
\item{data}{An array.}

\item{diminfo}{A list of information of the array.}
}
\description{
Create an expanded array with data and diminfo
}
