% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Calculating RMSE (root mean squared error)}
\usage{
rmse(obs, fit)
}
\arguments{
\item{obs}{A numeric vector of observed values.}

\item{fit}{A numeric vector of fitted values.}
}
\description{
Calculating RMSE (root mean squared error)
}
